requirements.txt
anyio==3.5.0
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
fastapi==0.75.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-decouple==3.6
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
